wormwood/main: Separate main from wormwood & add header for wormwood.
