+++ wormwood.h
@@ -1,0 +1,14 @@
+#pragma once
+
+void auth_user(void);
+
+void print_menu(void);
+
+void set_rod_depth(void);
+void set_flow_rate(void);
+
+void confirm_quit(void);
+
+void process_choice(char choice);
+
+void reactor_mgr_status(void);
